Reorganize deployments and add `abi` on each definition
