+++ src/abi/index.ts
@@ -1,0 +1,6 @@
+export { default as AzrealV0Abi } from './azrael.v0';
+export { default as ResolverV0Abi } from './resolver.v0';
+export { default as ResolverV1Abi } from './resolver.v1';
+export { default as SylvesterV0Abi } from './sylvester.v0';
+export { default as SylvesterV1Abi } from './sylvester.v1';
+export { default as WhoopiV0Abi } from './whoopi.v0';
